(Form/Fields) Rebase, refactor and reorganisation
