+++ src/components/Field/index.tsx
@@ -1,5 +1,3 @@
-export { default } from "./Field";
-export * from "./Field";
 export * from "./FieldTypes";
 export { sanitizeFieldSize } from "./fieldUtils";
