--- src/components/FieldWrapper/FieldWrapperTypes.ts
@@ -1,6 +1,0 @@
-import { ReactNode } from "react";
-
-export interface FieldWrapperProps {
-	children: ReactNode;
-	error?: boolean;
-}
