+++ src/components/FieldWrapper/index.tsx
@@ -1,1 +1,2 @@
 export { default } from "./FieldWrapper";
+export { default as HelperText } from "./HelperText";
